Call cuInit() in a separate thread, with a timeout
Call `cuInit()` in a separate thread, and fail if it does not manage to initialise the CUDA driver before a timeout.
This may happen e.g. if the CUDA driver is blocked by a running cuda-gdb.
Use the driver API rather than the runtime API because calls to `cudaGetErrorName()` and `cudaGetErrorString()` will
also block, while calls to `cuGetErrorName()` and `cuGetErrorString()` will work.