Implement simple MPI client and server for CMSSW
This implementation foresees two CMSSW processes running in parallel, on
the same or different machines:
  - a "driver" process;
  - a "follower" process.

The driver process is free to analyze any run/lumi/event from any source.
An "MPIDriver" module runs inside the driver process, listens to all
stream transitions (begin/end stream, begin/end run, begin/end lumi, new
event) and reads an arbitrary list of event data products.
The MPIDriver module connects over MPI to the follower process; it will
signal the same transitions to the follower process, and transfer to it
the event data products it reads.

The follower process uses a dedicated source, an "MPISource" module, to
recreate the same run, lumi and event transitions received from the
driver process.
The MPISource connects to the MPIDriver module in the driver process,
and will generate in the local process the same run and lumi transitions
and events that it receives; it will populate the local event with the
data product received from the driver, along with an additional product
used to keep track of the driver process rank and stream.

An automated test is available in the test/ directory.

Current limitations:
  - support a single stream
  - support a single "follower"
  - transfer only per-event data products