Reorganise the PortableCollection-related types
PortableHostCollection<T> is now a class template instead of a specialisation
of PortableDeviceCollection<T, Host>. As a result, it doesn't have any references
to Alpaka in its type name.

To avoid confusion and ODR-violations, the PortableDeviceCollection<T, TDev>
template can no longer be used with the Host device type.

The ALPAKA_ACCELERATOR_NAMESPACE::PortableCollection<T> is an alias template
that resolves to either PortableHostCollection<T> or
PortableDeviceCollection<T, ALPAKA_ACCELERATOR_NAMESPACE::Device>,
depending on the accelerator backend.

PortableCollection<T, TDev> is an alias template that resolves to
ALPAKA_ACCELERATOR_NAMESPACE::PortableCollection<T> for the matching
device.