Add support to multi layout portable collections (up to 5)
The design is recursive, but ROOT seems not to serialize variadic templates, so we have a
5 elements design. This number is arbitrary and can be changed with relatively little work.

The order of the PortableDeviceCollection template parameters is changed from device at the end
to device at the begining to accomodate the pseudo variadic templation.

A test validates the colleciton with two layouts.

The layout and view can be accessed by index or type, on the condition the type is present only
once in the collection.

The default index is set to 0 to minimize the need for chages in the single layout case. Empty
template parameters square bracket ("<>") are needed in some cases.