Extend CMSSW to a fully distributed application
Let multiple CMSSW processes on the same or different machines coordinate
event processing and transfer data products over MPI.

The implementation is based on four CMSSW modules.
Two are responsible for setting up the communication channels and
coordinate the event processing:
  - the MPIController
  - the MPISource
and two are responsible for the transfer of data products:
  - the MPISender
  - the MPIReceiver
.

The MPIController is an EDProducer running in a regular CMSSW process.
After setting up the communication with an MPISource, it transmits to it
all EDM run, lumi and event transitions, and instructs the MPISource to
replicate them in the second process.

The MPISource is a Source controlling the execution of a second CMSSW
process. After setting up the communication with an MPIController, it
listens for EDM run, lumi and event transitions, and replicates them in
its process.

Both MPIController and MPISource produce an MPIToken, a special data
product that encapsulates the information about the MPI communication
channel.

The MPISender is an EDProducer that can read a collection of a predefined
type from the Event, serialise it using its ROOT dictionary, and send it
over the MPI communication channel.

The MPIReceiver is an EDProducer that can receive a collection of a
predefined type over the MPI communication channel, deserialise is using
its ROOT dictionary, and put it in the Event.

Both MPISender and MPIReceiver are templated on the type to be
transmitted and de/serialised.

Each MPISender and MPIReceiver is configured with an instance value
that is used to match one MPISender in one process to one MPIReceiver in
another process. Using different instance values allows the use of
multiple MPISenders/MPIReceivers in a process.

Both MPISender and MPIReceiver obtain the MPI communication channel
reading an MPIToken from the event. They also produce a copy of the
MPIToken, so other modules can consume it to declare a dependency on
the previous modules.

An automated test is available in the test/ directory.