Make the caching memory allocator lock-free
Move the live block descriptors to the alpaka buffers: instead of tracking the
descriptors for the live blocks in a global map, store each block's descriptor
within the deleter of the block itself.

Reimplement the free list as a vector of tbb::concurrent_queue objects, one per
bin of the caching allocator.
Since a block may be in the free list but not actually be available for reuse,
blocks are popped from the queue until an available block is found and reused,
or the queue is empty and new block is requested. Then all popped blocks are
pushed back to the queue.

Use atomic operations for the individual statistics.
Access to the whole set of statistics may not be fully consistent, but they
should only be used for debugging or monitoring.