Implement a simple MPISender and MPIReceiver
The MPISender can read a collection of a predefined type from the Event,
serialise it using its ROOT dictionary, and send it over the MPI channel.

The MPIReceiver can receive a collection of a predefined type over the
MPI channel, deserialise is using its ROOT dictionary, and put it in the
Event.