Support generic types in MPISender/MPIReceiver
Let MPISender and MPIReceiver consume, send/receive and produce
collections of arbitrary types, as long as they have a ROOT dictionary
and can be persisted.

Note that any transient information is lost during the transfer, and
needs to be recreated by the receiving side.

The documentation and tests are updated accordingly.

Warning: this approach is a work in progress!
TODO:
  - improve framework integration
  - add checks between send/receive types