Fix PFClusterSoAProducer to read a device collection
Note: this is a quick workaround to let the device code use the device collection,
while being able to access the actual number of pf rechits on the host side.
It should replaced with a better and more general implementation, and the use of
the host collection should be removed.